+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add installer (It has not been officially released, but you can used it already).
-- Testing for single script file (Add `execute` evaluation).
+- Testing for single script file (Add `execute` evaluation, `Intercept` and `__SOURCED__` variable).
 
 ### Removed